// react-playlists/src/types/theAudioDB.d.ts
export interface TheAudioDBTrack {
  idTrack?: string;
  idArtist?: string;
  strTrack?: string;
  strArtist?: string;
  strAlbum?: string;
  strGenre?: string;
  strStyle?: string;
  strMood?: string;
  strTrackThumb?: string;
  strArtistThumb?: string;
  intYearReleased?: string;
  intFormedYear?: string;
}